pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "car-damage-detection"
version = "1.0.0"
description = "Deep learning toolkit for detecting and visualizing automotive exterior damage using YOLOv8"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["yolo", "object-detection", "computer-vision", "car-damage", "deep-learning", "pytorch"]
authors = [
    {name = "Car Damage Detection Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.2.95",
    "opencv-python>=4.10.0",
    "pillow>=10.4.0",
    "tkinterdnd2>=0.4.2",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "notebook>=6.4.0",
    "ipykernel>=6.0.0",
]
all = [
    "car-damage-detection[dev]",
]

[project.urls]
Homepage = "https://github.com/Oleksy1121/Car-damage-detection"
Repository = "https://github.com/Oleksy1121/Car-damage-detection"
Issues = "https://github.com/Oleksy1121/Car-damage-detection/issues"

[tool.setuptools]
packages = ["damage_detection_app"]
